+++ examples/console_logging_module.cpp
@@ -5,7 +5,9 @@
 
 import quill;
 
-#define QUILL_USE_MODULE
+#ifndef QUILL_USE_MODULE
+  #define QUILL_USE_MODULE
+#endif
 #include "quill/LogMacros.h"
 
 /**